Refactor Endpoint internals. It's ok to depend on `http` crate, and this reduces the number of additional abstractions needed.
