+++ Cargo.toml
@@ -27,15 +27,16 @@ tracing-span-filter = ["dep:tracing-subscriber"]
 bytes = "1.10"
 futures = "0.3"
 http = "1.3"
+http-body = "1.0.1"
 http-body-util = { version = "0.1", optional = true }
 hyper = { version = "1.6", optional = true}
 hyper-util = { version = "0.1", features = ["tokio", "server", "server-graceful", "http2"], optional = true }
 pin-project-lite = "0.2"
 rand = { version = "0.9", optional = true }
-regress = "0.10"
+regress = "=0.10.3"
 restate-sdk-macros = { version = "0.6", path = "macros" }
 restate-sdk-shared-core = { version = "=0.4.0", features = ["request_identity", "sha2_random_seed", "http"] }
-schemars = { version = "1.0.0-alpha.17", optional = true }
+schemars = { version = "1.0.0", optional = true }
 serde = "1.0"
 serde_json = "1.0"
 thiserror = "2.0"
